Added player

+++ Engine/Vec2.cpp
@@ -1,7 +1,7 @@
 #include "Vec2.h"
 #include <cmath>
 
-Vec2::Vec2( int x_in,int y_in )
+Vec2::Vec2( float x_in,float y_in )
 	:
 	x( x_in ),
 	y( y_in )
@@ -18,12 +18,12 @@ Vec2& Vec2::operator+=( const Vec2& rhs )
 	return *this = *this + rhs;
 }
 
-Vec2 Vec2::operator*( int rhs ) const
+Vec2 Vec2::operator*( float rhs ) const
 {
 	return Vec2( x * rhs,y * rhs );
 }
 
-Vec2& Vec2::operator*=( int rhs )
+Vec2& Vec2::operator*=( float rhs )
 {
 	return *this = *this * rhs;
 }
@@ -38,12 +38,12 @@ Vec2& Vec2::operator-=( const Vec2& rhs )
 	return *this = *this - rhs;
 }
 
-int Vec2::GetLength() const
+float Vec2::GetLength() const
 {
 	return std::sqrt( GetLengthSq() );
 }
 
-int Vec2::GetLengthSq() const
+float Vec2::GetLengthSq() const
 {
 	return x * x + y * y;
 }
@@ -55,7 +55,7 @@ Vec2& Vec2::Normalize()
 
 Vec2 Vec2::GetNormalized() const
 {
-	const int len = GetLength();
+	const float len = GetLength();
 	if( len != 0.0f )
 	{
 		return *this * (1.0f / len);
