+++ Engine/Vec2.h
@@ -4,18 +4,18 @@ class Vec2
 {
 public:
 	Vec2() = default;
-	Vec2( int x_in,int y_in );
+	Vec2( float x_in,float y_in );
 	Vec2 operator+( const Vec2& rhs ) const;
 	Vec2& operator+=( const Vec2& rhs );
-	Vec2 operator*( int rhs ) const;
-	Vec2& operator*=( int rhs );
+	Vec2 operator*( float rhs ) const;
+	Vec2& operator*=( float rhs );
 	Vec2 operator-( const Vec2& rhs ) const;
 	Vec2& operator-=( const Vec2& rhs );
-	int GetLength() const;
-	int GetLengthSq() const;
+	float GetLength() const;
+	float GetLengthSq() const;
 	Vec2& Normalize();
 	Vec2 GetNormalized() const;
 public:
-	int x;
-	int y;
+	float x;
+	float y;
 };
